+++ cli/src/merge_tools/external.rs
@@ -10,13 +10,14 @@ use itertools::Itertools;
 use jj_lib::backend::{FileId, MergedTreeId, TreeValue};
 use jj_lib::conflicts::{self, materialize_merge_result};
 use jj_lib::gitignore::GitIgnoreFile;
-use jj_lib::local_working_copy::{CheckoutError, SnapshotOptions, TreeState, TreeStateError};
+use jj_lib::local_working_copy::{CheckoutError, TreeState, TreeStateError};
 use jj_lib::matchers::Matcher;
 use jj_lib::merge::Merge;
 use jj_lib::merged_tree::{MergedTree, MergedTreeBuilder};
 use jj_lib::repo_path::RepoPath;
 use jj_lib::settings::UserSettings;
 use jj_lib::store::Store;
+use jj_lib::working_copy::SnapshotOptions;
 use regex::{Captures, Regex};
 use tempfile::TempDir;
 use thiserror::Error;
