+++ lib/tests/test_local_working_copy_concurrent.rs
@@ -16,9 +16,10 @@ use std::cmp::max;
 use std::thread;
 
 use assert_matches::assert_matches;
-use jj_lib::local_working_copy::{CheckoutError, SnapshotOptions};
+use jj_lib::local_working_copy::CheckoutError;
 use jj_lib::repo::Repo;
 use jj_lib::repo_path::RepoPath;
+use jj_lib::working_copy::{LockedWorkingCopy, SnapshotOptions};
 use jj_lib::workspace::Workspace;
 use testutils::{create_tree, write_working_copy_file, TestRepo, TestWorkspace};
